+++ src/endstone_runtime/bedrock/world/level/event/event_coordinator.cpp
@@ -160,13 +160,6 @@ void ServerInstanceEventCoordinator::sendServerThreadStopped(ServerInstance &ins
 void ServerInstanceEventCoordinator::sendServerLevelInitialized(ServerInstance &instance, Level &level)
 {
     auto &server = entt::locator<EndstoneServer>::value();
-    const static AutomaticID<Dimension, int> dimension_ids[] = {VanillaDimensions::Overworld, VanillaDimensions::Nether,
-                                                                VanillaDimensions::TheEnd};
-    for (const auto &dimension_id : dimension_ids) {
-        auto dimension = level.getOrCreateDimension(dimension_id);
-        auto endstone_level = std::make_unique<EndstoneLevel>(level, *dimension);
-        server.addLevel(std::move(endstone_level));
-    }
-
+    server.addLevel(std::make_unique<EndstoneLevel>(level));
     ENDSTONE_HOOK_CALL_ORIGINAL(&ServerInstanceEventCoordinator::sendServerLevelInitialized, this, instance, level);
 }